Adds a assigns clause module in contracts
